# Conflicts:
#	furms-core/src/main/java/io/imunity/furms/core/project_allocation/ProjectAllocationServiceValidator.java
#	furms-core/src/test/java/io/imunity/furms/core/project_allocation/ProjectAllocationServiceImplValidatorTest.java
#	furms-web-ui/src/main/java/io/imunity/furms/ui/views/community/projects/allocations/ProjectAllocationComponent.java

+++ furms-rabbitmq-site-client/src/main/java/io/imunity/furms/rabbitmq/site/client/ProjectAllocationMapper.java
@@ -18,7 +18,7 @@ class ProjectAllocationMapper {
 			.allocationIdentifier(projectAllocation.id)
 			.resourceCreditIdentifier(projectAllocation.resourceCredit.id)
 			.resourceType(projectAllocation.resourceType.name)
-			.amount(projectAllocation.amount.doubleValue())
+			.amount(projectAllocation.amount)
 			.validFrom(projectAllocation.resourceCredit.utcStartTime.atOffset(ZoneOffset.UTC))
 			.validTo(projectAllocation.resourceCredit.utcEndTime.atOffset(ZoneOffset.UTC))
 			.build();
